Change `qfox` to `pvdz` where relevant
The apparent switch of Github user account name from `qfox` to `pvdz`
broke some of the links in this repo. This commit fixes links of the
following forms:

    ...github.com/qfox/... --> ...github.com/pvdz/...
    ...//qfox.github.io/... --> ...//pvdz.github.io/...

I also took the liberty of changing standalone instances of `qFox` (such
as in the `README.md` title) to `PvdZ`, to clarify
ownership/maintainership.

`qfox` is left unchanged in the repo anywhere it has this context:

    ...qfox.nl/...

since that looks like the correct link.

This also bumps the version number from v0.0.4 to v0.0.5